Move tce server config back to config/server root
