--- config/server/common/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const tce = require('./tce');
-
-module.exports = { tce };
